+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This file documents the changes made to the formatter with each release.
 ### Fixed
 
 - Linter: Allow signal names that start with _ to be considered as snake_case (#356)
+- Fixed functions with multiple lines of arguments using method calls at the end of a chain with multiple object arguments resulting in a strange format (#349)
 
 ## Release 0.26.0 (2026-09-13)